Add files via upload
kali-anonstealth

ParrotSec's anonsurf and stealth, ported to work with Kali Linux.

How to use this repo

This repo contains the sources of both the anonsurf and pandora packages from ParrotSec combined into one.

Modifications have been made to use the DNS servers of Private Internet Access (instead of FrozenDNS), and fixes for users who don't use the resolvconf application. I have removed some functionality such as the gui and iceweasel in ram.

This repo can be compiled into a deb package to correctly install it on a Kali system.

The easiest way to get this working is to just run the installer. See the installation section for further info.

NOTE: This may work with any debian/ubuntu system, but this has only been tested to work on a kali-rolling amd64 system

Usage

Pandora

Pandora automatically overwrites the RAM when the system is shutting down. Pandora can also be ran manually:

pandora bomb
NOTE: This will clear the entire system cache, including active